refactor: optimize react-app config
