+++ packages/plugin-react-app/src/utils/index.js
@@ -1,0 +1,9 @@
+const getMiniappOutputPath = require('./getMiniappOutputPath');
+const getWebOutputPath = require('./getWebOutputPath');
+const logWebpackConfig = require('./logWebpackConfig');
+
+module.exports = {
+  getMiniappOutputPath,
+  getWebOutputPath,
+  logWebpackConfig
+};
